+++ NEWS.md
@@ -1,8 +1,17 @@
 # inlabru (development version)
 
+## Feaures
+  
+* Add `bru_get_mapper` generic, and associated methods for `inla.spde` and
+  `inla.rgeneric` objects. This allows `inlabru` to automatically extract
+  the appropriate `bru_mapper` object for each model component, and can be used
+  as a hook by external packages implementing new INLA object classes.
+
+## Bug fixes
+
 * Enable both datum/ensemble container for ellipsoid information, to support
   `epsg:4326`. Fixes #154
-
+  
 # inlabru 2.5.3
 
 ## Features